src/shoppingassistantservice/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --output-file=src/shoppingassistantservice/requirements.txt --pip-args=None src/shoppingassistantservice/requirements.in
#
aiofiles==24.1.0
    # via google-cloud-alloydb-connector
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via google-cloud-alloydb-connector
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via httpx
asyncpg==0.30.0
    # via
    #   google-cloud-alloydb-connector
    #   langchain-postgres
attrs==25.3.0
    # via aiohttp
blinker==1.9.0
    # via flask
cachetools==5.5.2
    # via google-auth
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via flask
colorama==0.4.6
    # via click
cryptography==45.0.7
    # via google-cloud-alloydb-connector
filetype==1.2.0
    # via langchain-google-genai
flask==3.1.2
    # via -r src/shoppingassistantservice/requirements.in
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
google-ai-generativelanguage==0.6.18
    # via langchain-google-genai
google-api-core[grpc]==2.25.1
    # via
    #   google-ai-generativelanguage
    #   google-cloud-alloydb
    #   google-cloud-alloydb-connector
    #   google-cloud-core
    #   google-cloud-secret-manager
    #   google-cloud-storage
google-auth==2.40.3
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-cloud-alloydb
    #   google-cloud-alloydb-connector
    #   google-cloud-core
    #   google-cloud-secret-manager
    #   google-cloud-storage
google-cloud-alloydb==0.4.9
    # via google-cloud-alloydb-connector
google-cloud-alloydb-connector[asyncpg]==1.9.0
    # via langchain-google-alloydb-pg
google-cloud-core==2.4.3
    # via google-cloud-storage
google-cloud-secret-manager==2.24.0
    # via -r src/shoppingassistantservice/requirements.in
google-cloud-storage==3.3.1
    # via langchain-google-alloydb-pg
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos[grpc]==1.70.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
greenlet==3.2.4
    # via sqlalchemy
grpc-google-iam-v1==0.14.2
    # via
    #   google-cloud-alloydb
    #   google-cloud-secret-manager
grpcio==1.74.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.74.0
    # via google-api-core
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langgraph-sdk
    #   langsmith
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.27
    # via -r src/shoppingassistantservice/requirements.in
langchain-core==0.3.75
    # via
    #   langchain
    #   langchain-google-genai
    #   langchain-postgres
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
langchain-google-alloydb-pg==0.13.0
    # via -r src/shoppingassistantservice/requirements.in
langchain-google-genai==2.1.10
    # via -r src/shoppingassistantservice/requirements.in
langchain-postgres==0.0.15
    # via langchain-google-alloydb-pg
langchain-text-splitters==0.3.11
    # via langchain
langgraph==0.2.60
    # via -r src/shoppingassistantservice/requirements.in
langgraph-checkpoint==2.1.1
    # via langgraph
langgraph-sdk==0.1.74
    # via langgraph
langsmith==0.4.26
    # via
    #   langchain
    #   langchain-core
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
multidict==6.6.4
    # via
    #   aiohttp
    #   yarl
numpy==2.3.2
    # via
    #   langchain-google-alloydb-pg
    #   langchain-postgres
    #   pgvector
orjson==3.11.3
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.10.0
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   langchain-core
    #   langsmith
pgvector==0.3.6
    # via langchain-postgres
pillow==11.3.0
    # via -r src/shoppingassistantservice/requirements.in
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-cloud-alloydb
    #   google-cloud-secret-manager
protobuf==6.32.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-cloud-alloydb
    #   google-cloud-alloydb-connector
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
psycopg[binary]==3.2.9
    # via
    #   -r src/shoppingassistantservice/requirements.in
    #   langchain-postgres
psycopg-binary==3.2.9
    # via psycopg
psycopg2-binary==2.9.10
    # via -r src/shoppingassistantservice/requirements.in
psycopg-pool==3.2.6
    # via langchain-postgres
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   langchain
    #   langchain-core
    #   langchain-google-genai
    #   langsmith
pydantic-core==2.33.2
    # via pydantic
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-core
requests==2.32.5
    # via
    #   google-api-core
    #   google-cloud-alloydb-connector
    #   google-cloud-storage
    #   langchain
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.43
    # via
    #   langchain
    #   langchain-postgres
tenacity==9.1.2
    # via langchain-core
typing-extensions==4.15.0
    # via
    #   langchain-core
    #   psycopg-pool
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via psycopg
urllib3==2.5.0
    # via requests
werkzeug==3.1.3
    # via flask
yarl==1.20.1
    # via aiohttp
zstandard==0.24.0
    # via langsmith
